+++ bindings/swift/Sources/LDKNode/LDKNode.swift
@@ -5080,10 +5080,11 @@ public struct Config {
     public var probingLiquidityLimitMultiplier: UInt64
     public var anchorChannelsConfig: AnchorChannelsConfig?
     public var routeParameters: RouteParametersConfig?
+    public var includeUntrustedPendingInSpendable: Bool
 
     // Default memberwise initializers are never public by default, so we
     // declare one manually.
-    public init(storageDirPath: String, network: Network, listeningAddresses: [SocketAddress]?, announcementAddresses: [SocketAddress]?, nodeAlias: NodeAlias?, trustedPeers0conf: [PublicKey], probingLiquidityLimitMultiplier: UInt64, anchorChannelsConfig: AnchorChannelsConfig?, routeParameters: RouteParametersConfig?) {
+    public init(storageDirPath: String, network: Network, listeningAddresses: [SocketAddress]?, announcementAddresses: [SocketAddress]?, nodeAlias: NodeAlias?, trustedPeers0conf: [PublicKey], probingLiquidityLimitMultiplier: UInt64, anchorChannelsConfig: AnchorChannelsConfig?, routeParameters: RouteParametersConfig?, includeUntrustedPendingInSpendable: Bool) {
         self.storageDirPath = storageDirPath
         self.network = network
         self.listeningAddresses = listeningAddresses
@@ -5093,6 +5094,7 @@ public struct Config {
         self.probingLiquidityLimitMultiplier = probingLiquidityLimitMultiplier
         self.anchorChannelsConfig = anchorChannelsConfig
         self.routeParameters = routeParameters
+        self.includeUntrustedPendingInSpendable = includeUntrustedPendingInSpendable
     }
 }
 
@@ -5127,6 +5129,9 @@ extension Config: Equatable, Hashable {
         if lhs.routeParameters != rhs.routeParameters {
             return false
         }
+        if lhs.includeUntrustedPendingInSpendable != rhs.includeUntrustedPendingInSpendable {
+            return false
+        }
         return true
     }
 
@@ -5140,6 +5145,7 @@ extension Config: Equatable, Hashable {
         hasher.combine(probingLiquidityLimitMultiplier)
         hasher.combine(anchorChannelsConfig)
         hasher.combine(routeParameters)
+        hasher.combine(includeUntrustedPendingInSpendable)
     }
 }
 
@@ -5151,16 +5157,17 @@ public struct FfiConverterTypeConfig: FfiConverterRustBuffer {
     public static func read(from buf: inout (data: Data, offset: Data.Index)) throws -> Config {
         return
             try Config(
-                storageDirPath: FfiConverterString.read(from: &buf), 
-                network: FfiConverterTypeNetwork.read(from: &buf), 
-                listeningAddresses: FfiConverterOptionSequenceTypeSocketAddress.read(from: &buf), 
-                announcementAddresses: FfiConverterOptionSequenceTypeSocketAddress.read(from: &buf), 
-                nodeAlias: FfiConverterOptionTypeNodeAlias.read(from: &buf), 
-                trustedPeers0conf: FfiConverterSequenceTypePublicKey.read(from: &buf), 
-                probingLiquidityLimitMultiplier: FfiConverterUInt64.read(from: &buf), 
-                anchorChannelsConfig: FfiConverterOptionTypeAnchorChannelsConfig.read(from: &buf), 
-                routeParameters: FfiConverterOptionTypeRouteParametersConfig.read(from: &buf)
-        )
+                storageDirPath: FfiConverterString.read(from: &buf),
+                network: FfiConverterTypeNetwork.read(from: &buf),
+                listeningAddresses: FfiConverterOptionSequenceTypeSocketAddress.read(from: &buf),
+                announcementAddresses: FfiConverterOptionSequenceTypeSocketAddress.read(from: &buf),
+                nodeAlias: FfiConverterOptionTypeNodeAlias.read(from: &buf),
+                trustedPeers0conf: FfiConverterSequenceTypePublicKey.read(from: &buf),
+                probingLiquidityLimitMultiplier: FfiConverterUInt64.read(from: &buf),
+                anchorChannelsConfig: FfiConverterOptionTypeAnchorChannelsConfig.read(from: &buf),
+                routeParameters: FfiConverterOptionTypeRouteParametersConfig.read(from: &buf),
+                includeUntrustedPendingInSpendable: FfiConverterBool.read(from: &buf)
+            )
     }
 
     public static func write(_ value: Config, into buf: inout [UInt8]) {
@@ -5173,6 +5180,7 @@ public struct FfiConverterTypeConfig: FfiConverterRustBuffer {
         FfiConverterUInt64.write(value.probingLiquidityLimitMultiplier, into: &buf)
         FfiConverterOptionTypeAnchorChannelsConfig.write(value.anchorChannelsConfig, into: &buf)
         FfiConverterOptionTypeRouteParametersConfig.write(value.routeParameters, into: &buf)
+        FfiConverterBool.write(value.includeUntrustedPendingInSpendable, into: &buf)
     }
 }